fix(task-4): cdk update, lambdas rearrange
