+++ product-service/src/handlers/getProductsList/app.ts
@@ -1,5 +1,5 @@
 import { buildResponse } from '../../utils/utils'
-import { products } from '../../mocks/data'
+import { getAllItems } from '../../db/dynamoDB/utils';
 import {
   APIGatewayProxyEvent,
   APIGatewayProxyResult
@@ -10,7 +10,24 @@ export const handler = async (
   try {
     console.log('event', event);
 
-    return buildResponse(200, products);
+    const [products, stocks] = await Promise.all([
+      getAllItems(process.env.TABLE_NAME_PRODUCT!),
+      getAllItems(process.env.TABLE_NAME_STOCK!)
+    ]);
+
+    if (!products || !stocks) {
+      return buildResponse(404, { message: 'Products are not found' });
+    }
+
+    const res = products.map((product) => {
+      const stock = stocks.find((stock) => stock.product_id === product.id);
+      return ({
+        ...product,
+        count: (stock?.count || 0)
+      });
+    });
+
+    return buildResponse(200, res);
   } catch (err: any) {
     return buildResponse(500, err.message)
   }
